+++ tests/conftest.py
@@ -40,17 +40,6 @@ def app_client(app: Flask) -> FlaskClient:
 
 
 @pytest.fixture
-def get_record_response(app_client: FlaskClient) -> ElementTree():
-    """Get a generated record for a given standard and config."""
-
-    def _get_record_response_for_config(standard: str, config: str) -> etree:
-        response = app_client.get(f"/standards/{standard}/{config}")
-        return fromstring(response.data)  # noqa: S320
-
-    return _get_record_response_for_config
-
-
-@pytest.fixture
 def fx_get_record_response(app_client: FlaskClient) -> ElementTree():
     """Get a generated record for a given standard/profile and config."""